+++ ts/packages/knowProTest/src/index.ts
@@ -6,3 +6,4 @@ export * from "./knowproContext.js";
 export * from "./searchTest.js";
 export * from "./answerTest.js";
 export * from "./memoryCommands.js";
+export * from "./podcastCommands.js";
